throw index exception for listpack insert and batch insert, add unit tests for get, getinteger with exception, refactor code of Set

+++ memory/intset.h
@@ -61,7 +61,7 @@ class IntSet::Iterator {
 
  private:
   const IntSet* intset_;
-  unsigned int idx_;
+  size_t idx_;
 };
 }  // namespace in_memory
 }  // namespace redis_simple
